requirements.txt
# pip install -r requirements.txt
flask
openai
python-dotenv
# pytest==6.2.5

